!refacto: refacto vendor system, serializelib
major refacto that try to better track the installation that happened and have more robust code.

We introduce a VendorInstallRecord file in the same sprit as the hub install file. And because it share the same code sign we introduce serializelib to have the same serialization code for those 2 files.